pyproject.toml
[project]
name = "merino-vss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.1",
    "sentence-transformers>=5.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "rich>=14.3.1",
    "scikit-learn>=1.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
